Add ContextWithAttrs support to golib/app2
